+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## staged
 
+- Add support for setbusxy dss command
 - Fix parsing of dss matrices and arrays with mixed delimiters (i.e., mixing commas and spaces)
 - Fix bug in dss parser to better support `object=` syntax
 - Add support for parsing of spectrum dss objects